Update 01_ruby/01_guess/play.rb
play.rb should call out the server when it has lied at least once (after being unable to guess the right number). It's unable to detect exactly when the server lied. 

The less efficient play.rb which checks every number incrementally from 1 to "limit" would be able to detect a pair of server responses, one of which must be a lie.